Update computed.md
Source of this revision: I found many incomprehensible places in the reading process. After careful reading, I found that the problem lies in the direct translation of English into Chinese. Many method names were directly translated into Chinese, which hindered the reading process and made it impossible to tell whether it was Chinese or the method in the function. So I revised the part that caused the misunderstanding.